backend: almost finished with hub

+++ server/quap/Models/GameModels/Player.cs
@@ -10,7 +10,7 @@ namespace quap.Models.GameModels
         public Guid PlayerId { get; set; }
         public string Name { get; set; }
         public string Icon { get; set; }
-        public float Score { get; set; } = 0;
+        public int Score { get; set; } = 0;
 
         public Guid GameId { get; set; }
         public Game Game { get; set; }
