+++ server/quap/Repositories/GameRepositories/IRepositories/IPlayerRepository.cs
@@ -8,7 +8,7 @@ namespace quap.Repositories.GameRepositories.IRepositories
     {
         Task<IEnumerable<Player>> GetAll(Guid gameId);
         Task<Player> CreatePlayer(CreatePlayerDto dto, Guid gameId);
-        Task<Player> AddScore(Guid playerId, float scoreToAdd);
+        Task<Player> AddScore(Guid playerId, int scoreToAdd);
         Task DeletePlayer(Guid playerId);
     }
 }
